requirements.txt
Django==2.2.2
django-pandas==0.6.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
joblib==0.13.2
numpy==1.16.4
pandas==0.24.2
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sqlparse==0.3.0
